requirements.txt
beautifulsoup4==4.13.3
pandas==2.2.3
python-dotenv==1.0.1
selenium==4.28.1
SQLAlchemy==2.0.38
mysql-connector-python==9.4.0
